+++ libvec/v3_dot.c
@@ -12,7 +12,7 @@
 
 #include "libvec.h"
 
-long double	v3_dot(t_vec3 a, t_vec3 b)
+REAL	v3_dot(t_vec3 a, t_vec3 b)
 {
 	return (a.x * b.x + a.y * b.y + a.z * b.z);
 }
